requirements.txt
Babel==2.3.4
click==6.6
Django==1.9.10
django-babel==0.5.1
django-braces==1.9.0
django-cors-headers==1.2.1
django-filter==0.15.2
django-grappelli==2.8.2
django-jinja==2.2.1
django-oauth-toolkit==0.10.0
djangorestframework==3.4.7
gevent==1.1.2
google-api-python-client==1.5.3
greenlet==0.4.10
hiredis==0.2.0
httplib2==0.9.2
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
oauth2client==3.0.0
oauthlib==1.0.3
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
pytz==2016.6.1
pyzmq==15.4.0
redis==2.10.5
rq==0.6.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
South==1.0.2
uritemplate==0.6
zmq==0.0.0
